:truck: 파일 이동 및 클래스 이름 변경
