--- src/main/java/com/kyungmin/lavanderia/jwt/util/TokenExpirationTime.java
@@ -1,6 +1,0 @@
-package com.kyungmin.lavanderia.jwt.util;
-
-public class TokenExpirationTime {
-    public static final Long ACCESS_TIME = 600000L;
-    public static final Long REFRESH_TIME = 86400000L;  // 86400L
-}
